+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/ExoPlayerTest.java
@@ -7197,42 +7197,6 @@ public class ExoPlayerTest {
   }
 
   @Test
-  public void prepare_preloadingEnabled_nextWindowPeriodCreatedForPreloading() throws Exception {
-    FakeMediaSource mediaSource1 =
-        new FakeMediaSource(
-            new FakeTimeline(
-                new TimelineWindowDefinition(
-                    /* isSeekable= */ true,
-                    /* isDynamic= */ false,
-                    /* durationUs= */ DefaultLoadControl.DEFAULT_MAX_BUFFER_MS * 2)));
-    List<MediaPeriodId> createdMediaPeriodIds = new ArrayList<>();
-    FakeMediaSource mediaSource2 =
-        new FakeMediaSource() {
-          @Override
-          public MediaPeriod createPeriod(
-              MediaPeriodId id, Allocator allocator, long startPositionUs) {
-            createdMediaPeriodIds.add(id);
-            return super.createPeriod(id, allocator, startPositionUs);
-          }
-        };
-    ExoPlayer player =
-        // Intentionally not using `parameterizeTestExoPlayerBuilder()` for preload specific test.
-        new TestExoPlayerBuilder(context)
-            .setPreloadConfiguration(
-                new ExoPlayer.PreloadConfiguration(/* targetPreloadDurationUs= */ 5_000_000L))
-            .build();
-    player.setMediaSources(ImmutableList.of(mediaSource1, mediaSource2));
-
-    player.prepare();
-    run(player).untilPendingCommandsAreFullyHandled();
-
-    assertThat(createdMediaPeriodIds).hasSize(1);
-    play(player).untilState(Player.STATE_ENDED);
-    assertThat(createdMediaPeriodIds).hasSize(1);
-    player.release();
-  }
-
-  @Test
   public void prepare_preloadingEnabledRepeatModeOne_sameWindowPeriodCreatedForPreloading()
       throws Exception {
     FakeTimeline timeline =
@@ -7243,7 +7207,8 @@ public class ExoPlayerTest {
                 /* durationUs= */ DefaultLoadControl.DEFAULT_MAX_BUFFER_MS * 2));
     List<MediaPeriodId> createdMediaPeriodIds = new ArrayList<>();
     FakeMediaSource mediaSource =
-        new FakeMediaSource(timeline) {
+        new FakeMediaSource(
+            timeline, ExoPlayerTestRunner.AUDIO_FORMAT, ExoPlayerTestRunner.VIDEO_FORMAT) {
           @Override
           public MediaPeriod createPeriod(
               MediaPeriodId id, Allocator allocator, long startPositionUs) {
@@ -7267,6 +7232,229 @@ public class ExoPlayerTest {
     player.setRepeatMode(Player.REPEAT_MODE_OFF);
     play(player).untilState(Player.STATE_ENDED);
     assertThat(createdMediaPeriodIds).hasSize(2);
+    player.release();
+  }
+
+  @Test
+  public void prepare_preloadingEnabled_nextWindowPeriodPreloaded() throws Exception {
+    List<MediaPeriodId> createdMediaPeriodIds = new ArrayList<>();
+    FakeMediaSource mediaSource1 =
+        new FakeMediaSource(
+            new FakeTimeline(
+                new TimelineWindowDefinition(
+                    /* isSeekable= */ true,
+                    /* isDynamic= */ false,
+                    /* durationUs= */ DefaultLoadControl.DEFAULT_MAX_BUFFER_MS * 2)),
+            ExoPlayerTestRunner.AUDIO_FORMAT,
+            ExoPlayerTestRunner.VIDEO_FORMAT) {
+          @Override
+          protected MediaPeriod createMediaPeriod(
+              MediaPeriodId id,
+              TrackGroupArray trackGroupArray,
+              Allocator allocator,
+              MediaSourceEventListener.EventDispatcher mediaSourceEventDispatcher,
+              DrmSessionManager drmSessionManager,
+              DrmSessionEventListener.EventDispatcher drmEventDispatcher,
+              @Nullable TransferListener transferListener) {
+            createdMediaPeriodIds.add(id);
+            return super.createMediaPeriod(
+                id,
+                trackGroupArray,
+                allocator,
+                mediaSourceEventDispatcher,
+                drmSessionManager,
+                drmEventDispatcher,
+                transferListener);
+          }
+        };
+    List<Long> preloadPreparationPositionUs = new ArrayList<>();
+    List<LoadingInfo> preloadLoadingInfos = new ArrayList<>();
+    FakeMediaSource mediaSource2 =
+        new FakeMediaSource(
+            new FakeTimeline(),
+            ExoPlayerTestRunner.AUDIO_FORMAT,
+            ExoPlayerTestRunner.VIDEO_FORMAT) {
+          @Override
+          protected MediaPeriod createMediaPeriod(
+              MediaPeriodId id,
+              TrackGroupArray trackGroupArray,
+              Allocator allocator,
+              MediaSourceEventListener.EventDispatcher mediaSourceEventDispatcher,
+              DrmSessionManager drmSessionManager,
+              DrmSessionEventListener.EventDispatcher drmEventDispatcher,
+              @Nullable TransferListener transferListener) {
+            createdMediaPeriodIds.add(id);
+            long positionInWindowUs =
+                getTimeline()
+                    .getPeriodByUid(id.periodUid, new Timeline.Period())
+                    .getPositionInWindowUs();
+            long defaultFirstSampleTimeUs = positionInWindowUs >= 0 ? 0 : -positionInWindowUs;
+            return new FakeMediaPeriod(
+                trackGroupArray,
+                allocator,
+                FakeMediaPeriod.TrackDataFactory.singleSampleWithTimeUs(defaultFirstSampleTimeUs),
+                mediaSourceEventDispatcher,
+                drmSessionManager,
+                drmEventDispatcher,
+                /* deferOnPrepared= */ false) {
+              @Override
+              public synchronized void prepare(Callback callback, long positionUs) {
+                preloadPreparationPositionUs.add(positionUs);
+                super.prepare(callback, positionUs);
+              }
+
+              @Override
+              public boolean continueLoading(LoadingInfo loadingInfo) {
+                preloadLoadingInfos.add(loadingInfo);
+                return super.continueLoading(loadingInfo);
+              }
+            };
+          }
+        };
+    MediaPeriodId firstMediaPeriodId =
+        new MediaPeriodId(/* periodUid= */ new Pair<>(0, 0), /* windowSequenceNumber= */ 0);
+    MediaPeriodId secondMediaPeriodId =
+        new MediaPeriodId(/* periodUid= */ new Pair<>(0, 0), /* windowSequenceNumber= */ 1);
+    ExoPlayer player =
+        // Intentionally not using `parameterizeTestExoPlayerBuilder()` for preload specific test.
+        new TestExoPlayerBuilder(context)
+            .setLoadControl(
+                new DefaultLoadControl() {
+                  @Override
+                  public boolean shouldContinuePreloading(
+                      Timeline timeline, MediaPeriodId mediaPeriodId, long bufferedDurationUs) {
+                    return true;
+                  }
+                })
+            .setPreloadConfiguration(
+                new ExoPlayer.PreloadConfiguration(/* targetPreloadDurationUs= */ 5_000_000L))
+            .build();
+    player.setMediaSources(ImmutableList.of(mediaSource1, mediaSource2));
+
+    player.prepare();
+    run(player).untilPendingCommandsAreFullyHandled();
+
+    // Assert both media periods are created, prepared and loaded when paused after preparation.
+    assertThat(createdMediaPeriodIds)
+        .containsExactly(firstMediaPeriodId, secondMediaPeriodId)
+        .inOrder();
+    assertThat(preloadPreparationPositionUs).containsExactly(123_000_000L);
+    assertThat(preloadLoadingInfos).hasSize(1);
+
+    play(player).untilState(Player.STATE_ENDED);
+
+    assertThat(createdMediaPeriodIds)
+        .containsExactly(firstMediaPeriodId, secondMediaPeriodId)
+        .inOrder();
+    // Verify that the preloaded period from the pool was used for enqueuing.
+    assertThat(preloadPreparationPositionUs).containsExactly(123_000_000L);
+    assertThat(preloadLoadingInfos).hasSize(1);
+    player.release();
+  }
+
+  @Test
+  public void prepare_preloadingDisabled_nextWindowPeriodNotPreloaded() throws Exception {
+    List<MediaPeriodId> createdMediaPeriodIds = new ArrayList<>();
+    FakeMediaSource mediaSource1 =
+        new FakeMediaSource(
+            new FakeTimeline(
+                new TimelineWindowDefinition(
+                    /* isSeekable= */ true,
+                    /* isDynamic= */ false,
+                    /* durationUs= */ DefaultLoadControl.DEFAULT_MAX_BUFFER_MS * 2)),
+            ExoPlayerTestRunner.AUDIO_FORMAT,
+            ExoPlayerTestRunner.VIDEO_FORMAT) {
+          @Override
+          protected MediaPeriod createMediaPeriod(
+              MediaPeriodId id,
+              TrackGroupArray trackGroupArray,
+              Allocator allocator,
+              MediaSourceEventListener.EventDispatcher mediaSourceEventDispatcher,
+              DrmSessionManager drmSessionManager,
+              DrmSessionEventListener.EventDispatcher drmEventDispatcher,
+              @Nullable TransferListener transferListener) {
+            createdMediaPeriodIds.add(id);
+            return super.createMediaPeriod(
+                id,
+                trackGroupArray,
+                allocator,
+                mediaSourceEventDispatcher,
+                drmSessionManager,
+                drmEventDispatcher,
+                transferListener);
+          }
+        };
+    List<Long> preloadPreparationPositionUs = new ArrayList<>();
+    List<LoadingInfo> preloadLoadingInfos = new ArrayList<>();
+    FakeMediaSource mediaSource2 =
+        new FakeMediaSource(
+            new FakeTimeline(),
+            ExoPlayerTestRunner.AUDIO_FORMAT,
+            ExoPlayerTestRunner.VIDEO_FORMAT) {
+          @Override
+          protected MediaPeriod createMediaPeriod(
+              MediaPeriodId id,
+              TrackGroupArray trackGroupArray,
+              Allocator allocator,
+              MediaSourceEventListener.EventDispatcher mediaSourceEventDispatcher,
+              DrmSessionManager drmSessionManager,
+              DrmSessionEventListener.EventDispatcher drmEventDispatcher,
+              @Nullable TransferListener transferListener) {
+            createdMediaPeriodIds.add(id);
+            long positionInWindowUs =
+                getTimeline()
+                    .getPeriodByUid(id.periodUid, new Timeline.Period())
+                    .getPositionInWindowUs();
+            long defaultFirstSampleTimeUs = positionInWindowUs >= 0 ? 0 : -positionInWindowUs;
+            return new FakeMediaPeriod(
+                trackGroupArray,
+                allocator,
+                FakeMediaPeriod.TrackDataFactory.singleSampleWithTimeUs(defaultFirstSampleTimeUs),
+                mediaSourceEventDispatcher,
+                drmSessionManager,
+                drmEventDispatcher,
+                /* deferOnPrepared= */ false) {
+              @Override
+              public synchronized void prepare(Callback callback, long positionUs) {
+                preloadPreparationPositionUs.add(positionUs);
+                super.prepare(callback, positionUs);
+              }
+
+              @Override
+              public boolean continueLoading(LoadingInfo loadingInfo) {
+                preloadLoadingInfos.add(loadingInfo);
+                return super.continueLoading(loadingInfo);
+              }
+            };
+          }
+        };
+    MediaPeriodId firstMediaPeriodId =
+        new MediaPeriodId(/* periodUid= */ new Pair<>(0, 0), /* windowSequenceNumber= */ 0);
+    MediaPeriodId secondMediaPeriodId =
+        new MediaPeriodId(/* periodUid= */ new Pair<>(0, 0), /* windowSequenceNumber= */ 1);
+    ExoPlayer player =
+        // Intentionally not using `parameterizeTestExoPlayerBuilder()` for preload specific test.
+        new TestExoPlayerBuilder(context)
+            .setPreloadConfiguration(ExoPlayer.PreloadConfiguration.DEFAULT)
+            .build();
+    player.setMediaSources(ImmutableList.of(mediaSource1, mediaSource2));
+
+    player.prepare();
+    run(player).untilPendingCommandsAreFullyHandled();
+
+    // Assert the media period of the second source isn't created yet.
+    assertThat(createdMediaPeriodIds).containsExactly(firstMediaPeriodId);
+    assertThat(preloadPreparationPositionUs).isEmpty();
+    assertThat(preloadLoadingInfos).isEmpty();
+
+    play(player).untilState(Player.STATE_ENDED);
+
+    // Assert the second second period is created for playback only.
+    assertThat(createdMediaPeriodIds)
+        .containsExactly(firstMediaPeriodId, secondMediaPeriodId)
+        .inOrder();
+    assertThat(preloadPreparationPositionUs).containsExactly(123_000_000L);
+    assertThat(preloadLoadingInfos).hasSize(1);
     player.release();
   }